pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "calibri"
version = "0.1.0"
description = "Official implementation of the paper Calibri: Enhancing Diffusion Transformers via Parameter-Efficient Calibration"
readme = "README.md"
requires-python = ">=3.10, <3.13"
license = { text = "MIT" }

dependencies = [
    "torch==2.5.1+cu118",
    "torchvision==0.20.1+cu118",

    "diffusers>=0.35.0",
    "accelerate>=1.10.0",
    "transformers==4.52.1",
    
    "absl-py>=2.3.0",
    "tensorboard>=2.20.0",
    "swanlab>=0.3.0",
    "tqdm>=4.67.0",
    
    "numpy>=2.0.0,<=2.2.6",
    "matplotlib>=3.10.0",
    "Pillow>=9.4.0",
    
    "cma>=4.4.0",
    "ml-collections>=1.1.0",
    "sentencepiece>=0.2.0",
    
    "image-reward>=1.5",
    "requests>=2.32.0",
    "openai>=1.0.0",
    
    "clip @ git+https://github.com/openai/CLIP.git"
]

[tool.uv]
package = false

# Указываем uv откуда скачивать PyTorch для CUDA 11.8
[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
torchaudio = { index = "pytorch-cu118" }
